Create Detector_FlatPanel.py
This callback function constructs a flat-panel detector. Specifically, it forces the number of module to be 1 and places the square-shaped flat panel perpendicular to the central x-ray, which simulates the equi-distance cone beam geometry. It also perform unform oversampling on the detector to approximate the pre-sampling image. When in use, it replaces the "Detector_ThirdGenCurved" in "scanner.detectorCallback"