Optical crosstalk of the flat panel detector
This callback function models the optical crosstalk of the flat panel detector. It applies a Lorentzian in frequency domain (which is characterized by a fitting parameter H) to the measured view after photon absorption and before electronic gain conversion. When in use, specify physics.opticalCrosstalkCallback = "Detection_OpticalCrosstalk_FlatPanel" in the cfg file